+++ kurtosis-module/impl/cl_client_network/cl_client_context.go
@@ -4,16 +4,17 @@ import "github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 
 type ConsensusLayerClientContext struct {
 	serviceCtx *services.ServiceContext
-	enode string
+	enr string
 }
 
-func NewConsensusLayerClientContext(serviceCtx *services.ServiceContext, enode string) *ConsensusLayerClientContext {
-	return &ConsensusLayerClientContext{serviceCtx: serviceCtx, enode: enode}
+func NewConsensusLayerClientContext(serviceCtx *services.ServiceContext, enr string) *ConsensusLayerClientContext {
+	return &ConsensusLayerClientContext{serviceCtx: serviceCtx, enr: enr}
 }
 
 func (ctx *ConsensusLayerClientContext) GetServiceContext() *services.ServiceContext {
 	return ctx.serviceCtx
 }
-func (ctx *ConsensusLayerClientContext) GetEnode() string {
-	return ctx.enode
+func (ctx *ConsensusLayerClientContext) GetENR() string {
+	return ctx.enr
 }
+