+++ docs/changelog.md
@@ -4,6 +4,8 @@
 * The Geth + Lighthouse node inside of Kurtosis now syncs with merge-devnet-3!
 * Successfully-working private, mining network!
 * Added a network of consensus-layer clients
+* Hooked up genesis generation for Geth & CL nodes
+* Lighthouse nodes peer with each other
 
 # 0.1.0
 * Initial commit